refactor: Moved logic out of main
