+++ siege-bot/src/main.rs
@@ -1,131 +1,29 @@
 mod commands;
 mod constants;
 pub mod formatting;
+pub mod handler;
 pub mod siege_player_lookup;
 
-use async_trait::async_trait;
 use serenity::{
-    model::{
-        gateway::Ready,
-        prelude::{interaction::Interaction, *},
-    },
-    prelude::{Context, EventHandler, RwLock, TypeMapKey},
+    model::prelude::*,
+    prelude::{RwLock, TypeMapKey},
     Client,
 };
 use siege_api::auth::Auth;
-use std::sync::Arc;
+use siege_player_lookup::SiegePlayerLookup;
+use std::{error::Error, sync::Arc};
 
 use tracing_subscriber::{fmt, prelude::*, EnvFilter};
 
-use commands::{all_maps::AllMapsCommand, CommandHandler};
-
-use crate::{
-    commands::{
-        add_player::AddPlayerCommand, all_operators::AllOperatorCommand, id::IdCommand,
-        map::MapCommand, operator::OperatorCommand, ping::PingCommand,
-        statistics::StatisticsCommand, CommandError,
-    },
-    siege_player_lookup::{PlayerLookupImpl, SiegePlayerLookup},
-};
-
-struct Handler;
-
-async fn sync_commands(guild_id: GuildId, ctx: &Context) {
-    tracing::info!("Syncing commands to {guild_id}");
-    match GuildId::set_application_commands(&guild_id, &ctx.http, |commands| {
-        commands
-            .create_application_command(|command| PingCommand::register(command))
-            .create_application_command(|command| IdCommand::register(command))
-            .create_application_command(|command| StatisticsCommand::register(command))
-            .create_application_command(|command| MapCommand::register(command))
-            .create_application_command(|command| OperatorCommand::register(command))
-            .create_application_command(|command| AddPlayerCommand::register(command))
-            .create_application_command(|command| AllOperatorCommand::register(command))
-            .create_application_command(|command| AllMapsCommand::register(command))
-    })
-    .await
-    {
-        Ok(commands) => tracing::trace!("Create guild slash commands: {commands:#?}"),
-        Err(err) => tracing::error!("Failed to create guild commands: {err:#?}"),
-    };
-}
-
-#[async_trait]
-impl EventHandler for Handler {
-    async fn ready(&self, _ctx: Context, ready: Ready) {
-        tracing::info!("Client is ready");
-        tracing::trace!("{ready:?}");
-
-        for guild in ready.guilds {
-            sync_commands(guild.id, &_ctx).await;
-        }
-    }
-
-    async fn guild_create(&self, ctx: Context, guild: Guild) {
-        tracing::info!("Connecting to guild: {:?}", guild.id);
-
-        let guild_id = guild.id;
-
-        GuildId::get_application_commands(&guild_id, &ctx.http)
-            .await
-            .iter()
-            .for_each(|id| tracing::debug!("Commands: {id:#?}"));
-    }
-
-    async fn interaction_create(&self, ctx: Context, interaction: Interaction) {
-        match interaction {
-            Interaction::ApplicationCommand(command) => {
-                tracing::trace!(
-                    "Received command interation from {guild_id:?}: {command:#?}",
-                    guild_id = command.guild_id
-                );
-
-                let result = match command.data.name.as_str() {
-                    "ping" => PingCommand::run(&ctx, &command).await,
-                    "id" => IdCommand::run(&ctx, &command).await,
-                    "statistics" => StatisticsCommand::run(&ctx, &command).await,
-                    "operator" => OperatorCommand::run(&ctx, &command).await,
-                    "map" => MapCommand::run(&ctx, &command).await,
-                    "add" => AddPlayerCommand::run(&ctx, &command).await,
-                    "all_operators" => AllOperatorCommand::run(&ctx, &command).await,
-                    "all_maps" => AllMapsCommand::run(&ctx, &command).await,
-                    _ => Err(CommandError::CommandNotFound),
-                };
-
-                if let Err(why) = result {
-                    tracing::error!("Failed to response to command: {why}");
-                }
-            }
-            Interaction::Autocomplete(autocomplete) => {
-                tracing::trace!("Autocomplete request: {autocomplete:#?}");
-
-                match autocomplete.data.name.as_str() {
-                    "operator" => {
-                        OperatorCommand::handle_autocomplete(&ctx, &autocomplete)
-                            .await
-                            .unwrap();
-                    }
-                    "map" => {
-                        MapCommand::handle_autocomplete(&ctx, &autocomplete)
-                            .await
-                            .unwrap();
-                    }
-                    name => tracing::warn!("Autocomplete for {name} not handled"),
-                }
-            }
-            _ => tracing::warn!("Unhanled interation: {interaction:?}"),
-        }
-    }
-}
+use crate::{handler::Handler, siege_player_lookup::PlayerLookupImpl};
 
 struct SiegeApi;
-
 impl TypeMapKey for SiegeApi {
     type Value = Arc<dyn siege_api::client::SiegeClient>;
 }
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
+async fn main() -> Result<(), Box<dyn Error>> {
     tracing_subscriber::registry()
         .with(fmt::layer())
         .with(
@@ -145,23 +43,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .await
         .expect("Error creating client");
 
+    setup_type_map(&mut client).await?;
+
     let shard_manager = client.shard_manager.clone();
-
-    // These extra scopes are added to drop the `GuardLock` on `data` as soon
-    // as possible. This not strictly necessary here, but in general best
-    // pratice to hold the locks as shortly as possible.
-    {
-        let siege_client: siege_api::client::Client =
-            Auth::from_environment().connect().await.unwrap().into();
-        let mut data = client.data.write().await;
-        data.insert::<SiegeApi>(Arc::new(siege_client));
-    }
-    {
-        let lookup = PlayerLookupImpl::load(".players.json")?;
-        let mut data = client.data.write().await;
-        data.insert::<SiegePlayerLookup>(Arc::new(RwLock::new(lookup)));
-    }
-
     tokio::spawn(async move {
         tokio::signal::ctrl_c()
             .await
@@ -175,6 +59,25 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // exponential backoff until it reconnects.
     if let Err(why) = client.start().await {
         println!("An error occurred while running the client: {:?}", why);
+    }
+
+    Ok(())
+}
+
+async fn setup_type_map(client: &mut Client) -> Result<(), Box<dyn Error>> {
+    // These extra scopes are added to drop the `GuardLock` on `data` as soon
+    // as possible. This not strictly necessary here, but in general best
+    // pratice to hold the locks as shortly as possible.
+    {
+        let siege_client: siege_api::client::Client =
+            Auth::from_environment().connect().await.unwrap().into();
+        let mut data = client.data.write().await;
+        data.insert::<SiegeApi>(Arc::new(siege_client));
+    }
+    {
+        let lookup = PlayerLookupImpl::load(".players.json")?;
+        let mut data = client.data.write().await;
+        data.insert::<SiegePlayerLookup>(Arc::new(RwLock::new(lookup)));
     }
 
     Ok(())
